Layouts/bottom navigation (#65)
* add actual folder

* merged and situated some files

---------
